# api/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "Dynamis Code Apps Template API",
    "version": "1.0.0",
    "description": "Stable v1 REST contract. Additive changes preserve authorization, ordering, errors, and side effects. Breaking changes require a new version or published migration window. Query collections use bounded filters, literal search, deterministic sorting, and opaque query-bound cursors. Errors use RFC 9457 Problem Details. See docs/api.md for the human-readable reference."
  },
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/members": {
      "get": {
        "operationId": "listMembers", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {"200": {"description": "Workspace members", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MemberPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/members/{userId}": {
      "patch": {
        "operationId": "changeMemberRole", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/UserId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/MemberRoleRequest"}}}},
        "responses": {"204": {"description": "Member role changed"}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      },
      "delete": {
        "operationId": "removeMember", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/UserId"}],
        "responses": {"204": {"description": "Member removed"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/ownership": {
      "post": {
        "operationId": "transferOwnership", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/OwnershipRequest"}}}},
        "responses": {"204": {"description": "Ownership transferred"}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations": {
      "get": {
        "operationId": "listInvitations", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {"200": {"description": "Workspace invitations", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      },
      "post": {
        "operationId": "createInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateInvitationRequest"}}}},
        "responses": {"201": {"description": "Invitation created", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationLinkResponse"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}/resend": {
      "post": {
        "operationId": "resendInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/InvitationId"}],
        "responses": {"200": {"description": "Invitation resent", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InvitationLinkResponse"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/invitations/{invitationId}": {
      "delete": {
        "operationId": "revokeInvitation", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/InvitationId"}],
        "responses": {"204": {"description": "Invitation revoked"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/tokens": {
      "get": {
        "operationId": "listTokens", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "responses": {"200": {"description": "Current user's workspace tokens", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      },
      "post": {
        "operationId": "createToken", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateTokenRequest"}}}},
        "responses": {"201": {"description": "Token created; secret is returned once", "headers": {"Cache-Control": {"schema": {"type": "string", "const": "no-store"}}}, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenSecretResponse"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/workspaces/{workspaceId}/tokens/{tokenId}": {
      "patch": {
        "operationId": "updateToken", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/TokenId"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenScopesRequest"}}}},
        "responses": {"204": {"description": "Token scopes changed"}, "400": {"$ref": "#/components/responses/Problem"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      },
      "delete": {
        "operationId": "revokeToken", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/WorkspaceId"}, {"$ref": "#/components/parameters/TokenId"}],
        "responses": {"204": {"description": "Token revoked"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/sessions": {
      "get": {
        "operationId": "listSessions", "security": [{"bearerAuth": []}],
        "responses": {"200": {"description": "Current user's sessions without secrets", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SessionPage"}}}}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v1/sessions/{sessionId}": {
      "delete": {
        "operationId": "revokeSession", "security": [{"bearerAuth": []}],
        "parameters": [{"$ref": "#/components/parameters/SessionId"}],
        "responses": {"204": {"description": "Session revoked"}, "401": {"$ref": "#/components/responses/Problem"}, "403": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer"},
      "sessionCookie": {"type": "apiKey", "in": "cookie", "name": "session"}
    },
    "parameters": {
      "WorkspaceId": {"name": "workspaceId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
...
      "UserId": {"name": "userId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
      "InvitationId": {"name": "invitationId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
      "TokenId": {"name": "tokenId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
      "SessionId": {"name": "sessionId", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[0-9a-f]{32}$"}},
...
    },
    "responses": {
      "Problem": {"description": "Problem Details", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}},
...
    },
    "schemas": {
...
      "Member": {"type": "object", "required": ["userId", "email", "role", "createdAt"], "properties": {"userId": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["owner", "admin", "member", "viewer"]}, "createdAt": {"type": "string", "format": "date-time"}}},
      "MemberPage": {"type": "object", "required": ["members"], "properties": {"members": {"type": "array", "items": {"$ref": "#/components/schemas/Member"}}}},
      "MemberRoleRequest": {"type": "object", "additionalProperties": false, "required": ["role"], "properties": {"role": {"type": "string", "enum": ["admin", "member", "viewer"]}}},
      "OwnershipRequest": {"type": "object", "additionalProperties": false, "required": ["userId"], "properties": {"userId": {"type": "string"}}},
      "Invitation": {"type": "object", "required": ["id", "workspaceId", "email", "role", "createdAt", "expiresAt"], "properties": {"id": {"type": "string"}, "workspaceId": {"type": "string"}, "email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["admin", "member", "viewer"]}, "createdAt": {"type": "string", "format": "date-time"}, "expiresAt": {"type": "string", "format": "date-time"}, "acceptedAt": {"type": ["string", "null"], "format": "date-time"}, "expiredAt": {"type": ["string", "null"], "format": "date-time"}, "revokedAt": {"type": ["string", "null"], "format": "date-time"}}},
      "InvitationPage": {"type": "object", "required": ["invitations"], "properties": {"invitations": {"type": "array", "items": {"$ref": "#/components/schemas/Invitation"}}}},
      "CreateInvitationRequest": {"type": "object", "additionalProperties": false, "required": ["email", "role"], "properties": {"email": {"type": "string", "format": "email"}, "role": {"type": "string", "enum": ["admin", "member", "viewer"]}, "lifetimeSeconds": {"type": "integer", "minimum": 60, "maximum": 2592000}}},
      "InvitationLinkResponse": {"type": "object", "required": ["invitation", "invitationUrl", "emailDelivered"], "properties": {"invitation": {"$ref": "#/components/schemas/Invitation"}, "invitationUrl": {"type": "string", "format": "uri-reference"}, "emailDelivered": {"type": "boolean"}}},
      "Token": {"type": "object", "required": ["id", "userId", "workspaceId", "name", "scopes", "createdAt"], "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "workspaceId": {"type": "string"}, "name": {"type": "string"}, "scopes": {"type": "array", "items": {"type": "string"}}, "createdAt": {"type": "string", "format": "date-time"}, "expiresAt": {"type": ["string", "null"], "format": "date-time"}, "lastUsedAt": {"type": ["string", "null"], "format": "date-time"}, "revokedAt": {"type": ["string", "null"], "format": "date-time"}}},
      "TokenPage": {"type": "object", "required": ["tokens"], "properties": {"tokens": {"type": "array", "items": {"$ref": "#/components/schemas/Token"}}}},
      "CreateTokenRequest": {"type": "object", "additionalProperties": false, "required": ["name", "scopes"], "properties": {"name": {"type": "string", "minLength": 1, "maxLength": 100}, "scopes": {"type": "array", "minItems": 1, "items": {"type": "string"}}, "expiresAt": {"type": ["string", "null"], "format": "date-time"}}},
      "TokenScopesRequest": {"type": "object", "additionalProperties": false, "required": ["scopes"], "properties": {"scopes": {"type": "array", "minItems": 1, "items": {"type": "string"}}}},
      "TokenSecretResponse": {"type": "object", "required": ["token", "secret"], "properties": {"token": {"$ref": "#/components/schemas/Token"}, "secret": {"type": "string"}}},
      "Session": {"type": "object", "required": ["id", "userId", "authMethod", "authLevel", "createdAt", "expiresAt"], "properties": {"id": {"type": "string"}, "userId": {"type": "string"}, "authMethod": {"type": "string"}, "authLevel": {"type": "integer", "minimum": 1, "maximum": 2}, "oidcProviderId": {"type": "string"}, "createdAt": {"type": "string", "format": "date-time"}, "expiresAt": {"type": "string", "format": "date-time"}, "revokedAt": {"type": ["string", "null"], "format": "date-time"}}},
      "SessionPage": {"type": "object", "required": ["sessions"], "properties": {"sessions": {"type": "array", "items": {"$ref": "#/components/schemas/Session"}}}},
...
      "Problem": {"type": "object", "required": ["type", "title", "status", "detail", "instance", "code", "requestId"], "properties": {"type": {"type": "string", "format": "uri"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}, "instance": {"type": "string"}, "code": {"type": "string"}, "requestId": {"type": "string"}}}
    }
  }
}